+++ docx-core/src/documents/elements/table_width.rs
@@ -20,7 +20,7 @@ impl TableWidth {
 impl BuildXML for TableWidth {
     fn build(&self) -> Vec<u8> {
         XMLBuilder::new()
-            .table_width(self.width, WidthType::DXA)
+            .table_width(self.width as i32, WidthType::DXA)
             .build()
     }
 }
